Перенос в отдельную библиотеку
