+++ WordCloudImageGenerator/Layouting/Layouters/LayoutTypes.cs
@@ -1,4 +1,4 @@
-﻿namespace WordCloud.WordCloudRenedering
+namespace WordCloudImageGenerator.LayoutCraetion.Layouters
 {
     public enum LayoutTypes
     {
