+++ WordCloudImageGenerator/Parsing/Extractors/IWordExtractor.cs
@@ -1,10 +1,6 @@
-﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace WordCloud.TextAnalyze.Extractors
+namespace WordCloudImageGenerator.Parsing.Extractors
 {
     public interface IWordExtractor
     {
